App for projection generation and emulate_print.m work
Created app for projection generation which simplifies the creation of the projection set cell array. Settings can be changed with immediate visual feedback and then applied to all projections. After this the projection set is created and the user can move to the next window for projector control (work in progress).

Work on emulate_print.m